+++ src/components/Cart.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import formatPrice from '../helpers/formatPrice'
+import formatPrice from '../helpers/formatPrice';
 
-const Cart = ({ items, subTotal, tax, total }) => {
+Cart = ({ items, subTotal, tax, total }) => {
     const listItems = items.map(item=>{
         return (
             <li key={item.id}>{item.name}: ${item.price}</li>
